requirements.txt
dlib==19.24.0
gradio==3.4.1
imutils==0.5.4
matplotlib
numpy
opencv_python==4.6.0.66
Pillow==9.2.0
scikit-image
scipy
torch==1.12.1
tqdm
